requirements.txt
python-dotenv
slack
gputil
flask
slack-sdk
requests
pytest